requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
dnspython==2.6.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pymongo==4.7.2
python-dotenv==1.0.1
sqlparse==0.5.0
